+++ src/server/physics/event/ig_ENTITY_EventTrigger.ts
@@ -3,17 +3,34 @@ import { isPhysics } from '../is-physics-server'
 import { setEventNextTriggeredBy, unsetEventNextTriggeredBy } from './event-manager'
 import { findSetByEntityByVars, setNextSetBy, unsetNextSetBy } from './vars'
 
+declare global {
+    namespace ig.ENTITY {
+        namespace EventTrigger {
+            interface Settings {
+                forceRunOnMap?: boolean
+            }
+        }
+        interface EventTrigger {
+            forceRunOnMap?: boolean
+        }
+    }
+}
+
 prestart(() => {
     if (!PHYSICS) return
 
     ig.ENTITY.EventTrigger.inject({
+        init(x, y, z, settings) {
+            this.parent(x, y, z, settings)
+            this.forceRunOnMap = settings.forceRunOnMap
+        },
         update() {
             if (!isPhysics(multi.server)) return this.parent()
 
             const setBy = findSetByEntityByVars(this.startCondition.vars)
             if (setBy) setNextSetBy(setBy)
 
-            setEventNextTriggeredBy(this.startCondition)
+            if (!this.forceRunOnMap) setEventNextTriggeredBy(this.startCondition)
 
             this.parent()